Added functions to create the different columns of the matrix
the words of the column are not case sensitive and do not contain punctuation. They may contain numbers (not sure if this will be useful, maybe for years. not sure). Next is to add rows to fill in the entries. I added a function generateCounts that must be filled in to do this